Commit of previously released v1.0
The project was released on userstyles.org without having any source
control. Although version 1.0 was already released on 05/16/2015 this
commit is just now initializing a git repository. Version 1.0 as
previously released lacks files currently present within this commit.
The contents of this commit are considered authoritative and are all
considered to be initialized as part of version 1.0. Any modifications
of any project files after this commit will be a part of a later
version (this is an exceptional case and not all future modifications
should increment the version).

This project is licensed under CC BY-NC-SA; see the license.md file for
the full details.

Add "Additional Info.txt" - the additional info section of the
    userstyles.org page
Add "Description.txt" - the description section of the userstyles.org
    page
Add "YLO YTC Fix Screenshot.jpg" - the main "results" screenshot of the
    userstyles.org page
Add "gitmessage.txt" - the default commit template with basic style
    rules included for the benefit of collaborators
Add "license.md" - the Creative Commons license in Markdown format
Add "readme.md" - the GitHub parsed README in Markdown format
Add "userstyle.css" - the CSS of the userstyle that is downloaded by the
    end-user